拆分core的生成部分
